Rename JSONSerializer filename to match constant
