+++ lib/vault/rails/json_serializer.rb
